requirements.txt
pandas
google-cloud-bigquery
playwright